Prepare codebase for dynamic object creation
Used inheritance to make all objects of the same type so that global objects can be declared, but then those objects can have different subclass types. This is needed so that calls to the port objects can be type agnostic. Ports will be objects, and their type will change depending on what is connected to them and the configuration options selected in the menu system.